+++ changes.md
@@ -4,3 +4,5 @@
 
 - When requesting to remove stream configurations, wildcards '*' and '?' can now be
 used in the output topic and channel name fields.
+
+- Can request Forwarder to store the current configuration and retrieve it when restarting.